refactor: reduce file sizes to meet CI limits
- git_client.py: 320->254 lines (merge imports, extract check_merge_conflicts)
- serve.py: 304->231 lines (move _build_server to serve_utils.py)
- serve_utils.py: receive _build_server function
- git_branch_ops.py: add check_merge_conflicts function
- add plan-phase2-governance-circuit-breaker.md for deferred items
